requirements.txt
# Direct dependencies only (install with: pip install -r requirements.txt)
streamlit>=1.40
pdfplumber>=0.11
python-docx>=1.1
google-genai>=1.0
pydantic>=2.0
fpdf2>=2.7
python-dotenv>=1.0
